# services/route-service/internal/api/v1/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API serves as an interface to interact with the My Route Service platform, providing endpoints for managing bus routes, Stops, and schedules.",
        "title": "My Route Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Mohammad Mizan",
            "url": "http://swagger.io/support",
            "email": "takbir.jcd@gmail.com"
        },
        "license": {
            "name": "Apache License Version 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/routes": {
            "get": {
                "description": "Get a list of all routes available in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Retrieve all routes",
                "responses": {
                    "200": {
                        "description": "List of all routes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.RouteInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Unable to retrieve routes",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new route with the specified details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Add a new route",
                "parameters": [
                    {
                        "description": "Create Route Request",
                        "name": "route",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RouteCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created route information",
                        "schema": {
                            "$ref": "#/definitions/dto.RouteInfo"
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Unable to create route",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/routes/{routeId}": {
            "get": {
                "description": "Retrieve a route by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Get route by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved route",
                        "schema": {
                            "$ref": "#/definitions/dto.RouteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid route ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "404": {
                        "description": "Route not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Unable to retrieve route",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            },
            "put": {
                "description": "Update the details of an existing route by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Update a route",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "uint",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Route Update Information",
                        "name": "route",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RouteUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated route information",
                        "schema": {
                            "$ref": "#/definitions/dto.RouteInfo"
                        }
                    },
                    "400": {
                        "description": "Invalid route ID or request format",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Unable to update route",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a route by its unique ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "routes"
                ],
                "summary": "Delete a route",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Route successfully deleted"
                    },
                    "400": {
                        "description": "Invalid route ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    },
                    "500": {
                        "description": "Unable to delete route",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorMessage"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.BusResponse": {
            "type": "object",
            "properties": {
                "busCode": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lastServiceDate": {
                    "type": "string"
                },
                "licensePlate": {
                    "type": "string"
                },
                "nextServiceDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "dto.RouteCreateRequest": {
            "type": "object",
            "required": [
                "endLocation",
                "name",
                "startLocation"
            ],
            "properties": {
                "endLocation": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "startLocation": {
                    "type": "string"
                }
            }
        },
        "dto.RouteInfo": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "end_location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "route_id": {
                    "type": "integer"
                },
                "start_location": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.RouteResponse": {
            "type": "object",
            "properties": {
                "buses": {
                    "description": "Optional Buses",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.BusResponse"
                    }
                },
                "endLocation": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "startLocation": {
                    "type": "string"
                },
                "stops": {
                    "description": "Optional Stops",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StopResponse"
                    }
                }
            }
        },
        "dto.RouteUpdateRequest": {
            "type": "object",
            "properties": {
                "endLocation": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "startLocation": {
                    "type": "string"
                }
            }
        },
        "dto.ScheduleResponse": {
            "type": "object",
            "properties": {
                "arrival_time": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "departure_time": {
                    "type": "string"
                },
                "schedule_id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.StopResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ScheduleResponse"
                    }
                },
                "sequence": {
                    "type": "integer"
                },
                "stop_id": {
                    "type": "integer"
                }
            }
        },
...
        "pkg.ErrorMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}